tests: break to small tests
