+++ example/integration_test/constant.dart
@@ -1,0 +1,26 @@
+const String appVersion = "1.2.3";
+const String oldAppVersion = "0.0.1";
+const bool debugging = true;
+
+// E2E Flutter
+const String featureTag = "flutter";
+const String userId = 'bucketeer-flutter-user-id-1';
+
+const String featureIdBoolean = "feature-flutter-e2e-boolean";
+const bool featureIdBooleanValue = true;
+
+const String featureIdString = "feature-flutter-e2e-string";
+const String featureIdStringValue = "value-1";
+const String featureIdStringValueUpdate = "value-2";
+
+const String featureIdInt = "feature-flutter-e2e-int";
+const int featureIdIntValue = 10;
+
+const String featureIdDouble = "feature-flutter-e2e-double";
+const double featureIdDoubleValue = 2.1;
+
+const String featureIdJson = "feature-flutter-e2e-json";
+const Map<String, dynamic> featureIdJsonValue = {"key": "value-1"};
+
+const String goalId = "goal-flutter-e2e-1";
+const double goalValue = 1.0;
