+++ example/integration_test/main_usage_test.dart
@@ -6,39 +6,11 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:bucketeer_flutter_client_sdk/bucketeer_flutter_client_sdk.dart';
 import 'package:mocktail/mocktail.dart';
-
-class MockEvaluationUpdateListener extends Mock
-    implements BKTEvaluationUpdateListener {}
+import 'constant.dart';
+import 'helper.dart';
 
 void main() async {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
-
-  const String appVersion = "1.2.3";
-  const String oldAppVersion = "0.0.1";
-  const bool debugging = true;
-
-  // E2E Flutter
-  const String featureTag = "flutter";
-  const String userId = 'bucketeer-flutter-user-id-1';
-
-  const String featureIdBoolean = "feature-flutter-e2e-boolean";
-  const bool featureIdBooleanValue = true;
-
-  const String featureIdString = "feature-flutter-e2e-string";
-  const String featureIdStringValue = "value-1";
-  const String featureIdStringValueUpdate = "value-2";
-
-  const String featureIdInt = "feature-flutter-e2e-int";
-  const int featureIdIntValue = 10;
-
-  const String featureIdDouble = "feature-flutter-e2e-double";
-  const double featureIdDoubleValue = 2.1;
-
-  const String featureIdJson = "feature-flutter-e2e-json";
-  const Map<String, dynamic> featureIdJsonValue = {"key": "value-1"};
-
-  const String goalId = "goal-flutter-e2e-1";
-  const double goalValue = 1.0;
 
   group('Bucketeer: general test', () {
     setUpAll(() async {
@@ -430,196 +402,5 @@ void main() async {
       }).onError((error, stackTrace) => fail(
           "destroy() should success and should not throw an exception ${error.toString()}"));
     });
-  });
-
-  group('Bucketeer: test optional configurations', () {
-    setUp(() {});
-
-    test('BKTClient should allow feature_tag to be optional', () async {
-      final config = BKTConfigBuilder()
-          .apiKey(Constants.apiKey)
-          .apiEndpoint(Constants.apiEndpoint)
-          .debugging(debugging)
-          .eventsMaxQueueSize(Constants.exampleEventMaxQueueSize)
-          .eventsFlushInterval(Constants.exampleEventsFlushInterval)
-          .pollingInterval(Constants.examplePollingInterval)
-          .backgroundPollingInterval(Constants.exampleBackgroundPollingInterval)
-          .appVersion(appVersion)
-          .build();
-      assert(config.featureTag == "");
-      final user = BKTUserBuilder().id(userId).customAttributes({}).build();
-
-      await E2EBKTClient.initializeWithRetryMechanism(
-        config: config,
-        user: user,
-      ).then((instanceResult) {
-        expect(instanceResult.isInitializeSuccess(), true,
-            reason: "initialize() should success");
-      }, onError: (obj, st) {
-        fail('initialize() should not throw an exception');
-      });
-
-      /// init without feature tag should retrieves all features
-      final android = await BKTClient.instance
-          .evaluationDetails("feature-android-e2e-string");
-      expect(android != null, true,
-          reason: "evaluationDetails should not be null");
-
-      final golang =
-          await BKTClient.instance.evaluationDetails("feature-go-server-e2e-1");
-      expect(golang != null, true,
-          reason: "evaluationDetails should not be null");
-
-      final javascript =
-          await BKTClient.instance.evaluationDetails("feature-js-e2e-string");
-      expect(javascript != null, true,
-          reason: "evaluationDetails should not be null");
-    });
-
-    tearDown(() async {
-      await BKTClient.instance.destroy().then((value) {
-        expect(value.isSuccess, true, reason: "destroy() should success");
-      }, onError: (obj, st) {
-        fail("destroy() should success and should not throw an exception");
-      });
-    });
-  });
-
-  group('Bucketeer: error handling', () {
-    testWidgets('access BKTClient before initializing', (WidgetTester _) async {
-      var completer = Completer<BKTResult<void>>();
-      BKTResult<void> fetchEvaluationsRs =
-          await BKTClient.instance.fetchEvaluations().then((value) {
-        /// Use completer to make sure this callback will get called
-        /// even if BKTClient has not initialize
-        completer.complete(value);
-        return value;
-      }, onError: (obj, st) {
-        fail("fetchEvaluations() should not throw an exception");
-      });
-      expect(fetchEvaluationsRs.isFailure, true,
-          reason:
-              "fetchEvaluations() should fail ${fetchEvaluationsRs.toString()}");
-      expect(fetchEvaluationsRs.asFailure.exception,
-          isA<BKTIllegalStateException>(),
-          reason:
-              "exception should be BKTIllegalStateException but got ${fetchEvaluationsRs.toString()}");
-
-      BKTResult<void> flushRs = await BKTClient.instance.flush().then((value) {
-        return value;
-      }, onError: (obj, st) {
-        fail("flush() should not throw an exception");
-      });
-      expect(flushRs.isFailure, true, reason: "flush() should fail");
-      expect(fetchEvaluationsRs.asFailure.exception,
-          isA<BKTIllegalStateException>(),
-          reason:
-              "exception should be BKTIllegalStateException but got ${fetchEvaluationsRs.toString()}");
-
-      expect(completer.isCompleted, true,
-          reason: "completer should be completed");
-    });
-
-    testWidgets('initialize BKTClient with invalid API_KEY',
-        (WidgetTester _) async {
-      final config = BKTConfigBuilder()
-          .apiKey("RANDOM_KEY")
-          .apiEndpoint(Constants.apiEndpoint)
-          .debugging(debugging)
-          .eventsMaxQueueSize(Constants.exampleEventMaxQueueSize)
-          .eventsFlushInterval(Constants.exampleEventsFlushInterval)
-          .pollingInterval(Constants.examplePollingInterval)
-          .backgroundPollingInterval(Constants.exampleBackgroundPollingInterval)
-          .appVersion(appVersion)
-          .build();
-      assert(config.featureTag == "");
-      final user = BKTUserBuilder().id(userId).customAttributes({}).build();
-
-      await E2EBKTClient.initializeWithRetryMechanism(
-        config: config,
-        user: user,
-      ).then((instanceResult) {
-        expect(instanceResult.isFailure, true,
-            reason: "initialize() should fail ${instanceResult.toString()}");
-        expect(instanceResult.asFailure.exception, isA<BKTForbiddenException>(),
-            reason:
-                "exception should be BKTForbiddenException but got ${instanceResult.toString()}. The exception could be a BKTTimeoutException, but we don't want it here");
-      }, onError: (obj, st) {
-        fail("initialize() should not throw an exception");
-      });
-
-      await BKTClient.instance.fetchEvaluations().then((fetchEvaluationsRs) {
-        expect(fetchEvaluationsRs.asFailure.exception,
-            isA<BKTForbiddenException>(),
-            reason:
-                "exception should be BKTForbiddenException but got ${fetchEvaluationsRs.toString()}");
-      }, onError: (obj, st) {
-        fail(
-            "fetchEvaluations() should not throw an exception ${obj.toString()}");
-      });
-
-      await BKTClient.instance.flush().then((flushRs) {
-        expect(flushRs.asFailure.exception, isA<BKTForbiddenException>(),
-            reason:
-                "exception should be BKTForbiddenException but got ${flushRs.toString()}");
-        return flushRs;
-      }, onError: (obj, st) {
-        fail(
-            "fetchEvaluations() should not throw an exception but got ${obj.toString()}");
-      });
-
-      await BKTClient.instance.destroy().then(
-          (value) =>
-              expect(value.isSuccess, true, reason: "destroy() should succeed"),
-          onError: (obj, st) {
-        fail("destroy() should not throw an exception");
-      });
-
-      await BKTClient.instance.fetchEvaluations().then((fetchEvaluationsRs) {
-        expect(fetchEvaluationsRs.isFailure, true,
-            reason: "fetchEvaluations() should fail");
-        expect(fetchEvaluationsRs.asFailure.exception,
-            isA<BKTIllegalStateException>(),
-            reason:
-                "exception should be BKTIllegalStateException but got ${fetchEvaluationsRs.toString()}");
-      }, onError: (obj, st) {
-        fail(
-            "fetchEvaluations() should not throw an exception ${obj.toString()}");
-      });
-    });
-  });
-}
-
-extension InitializeSuccess on BKTResult<void> {
-  bool isInitializeSuccess() {
-    return isSuccess || asFailure.exception is BKTTimeoutException;
-  }
-}
-
-extension E2EBKTClient on BKTClient {
-  /// This func will retry the BKTClient.initialize 3 times (default) if we got a network error.
-  static Future<BKTResult<void>> initializeWithRetryMechanism({
-    required BKTConfig config,
-    required BKTUser user,
-    int? timeoutMillis,
-    int maxRetryTimes = 3,
-  }) async {
-    var retryCount = 0;
-    BKTResult<void>? result;
-    do {
-      result = await BKTClient.initialize(
-        config: config,
-        user: user,
-        timeoutMillis: timeoutMillis,
-      );
-      bool shouldRetry =
-          result.isFailure && result.asFailure is BKTNetworkException;
-      if (shouldRetry) {
-        retryCount++;
-        continue;
-      }
-      break;
-    } while (retryCount < maxRetryTimes);
-    return result;
-  }
+  }, skip: false);
 }
